requirements.txt
fastapi
uvicorn
PyPDF2
openai
python-dotenv